requirements.txt
alembic==1.11.1
greenlet==2.0.2
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.5.0
tzdata==2023.3
